+++ bootique-agrest/src/main/java/io/bootique/agrest/AgrestModuleExtender.java
@@ -32,7 +32,7 @@ public class AgrestModuleExtender extends ModuleExtender<AgrestModuleExtender> {
 
     private Multibinder<AgFeatureProvider> featureProviders;
     private Multibinder<AgModuleProvider> moduleProviders;
-
+    private Multibinder<AgBuilderCallback> builderCallbacks;
 
     public AgrestModuleExtender(Binder binder) {
         super(binder);
@@ -42,6 +42,25 @@ public class AgrestModuleExtender extends ModuleExtender<AgrestModuleExtender> {
     public AgrestModuleExtender initAllExtensions() {
         contributeFeatureProviders();
         contributeModuleProviders();
+        contributeBuilderCallbacks();
+        return this;
+    }
+
+    /**
+     * @return this extender instance.
+     * @since 1.1
+     */
+    public AgrestModuleExtender addBuilderCallback(AgBuilderCallback callback) {
+        contributeBuilderCallbacks().addBinding().toInstance(callback);
+        return this;
+    }
+
+    /**
+     * @return this extender instance.
+     * @since 1.1
+     */
+    public AgrestModuleExtender addBuilderCallback(Class<? extends AgBuilderCallback> callbackType) {
+        contributeBuilderCallbacks().addBinding().to(callbackType);
         return this;
     }
 
@@ -91,5 +110,9 @@ public class AgrestModuleExtender extends ModuleExtender<AgrestModuleExtender> {
 
     private Multibinder<AgModuleProvider> contributeModuleProviders() {
         return moduleProviders != null ? moduleProviders : (moduleProviders = newSet(AgModuleProvider.class));
+    }
+
+    private Multibinder<AgBuilderCallback> contributeBuilderCallbacks() {
+        return builderCallbacks != null ? builderCallbacks : (builderCallbacks = newSet(AgBuilderCallback.class));
     }
 }